backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
python-dotenv
pandas
scikit-learn
joblib
reportlab
qrcode
bcrypt
